Minor additions and FAG stuff.

+++ d3x/Structures/ObjectData.cs
@@ -9,8 +9,8 @@ namespace D3x.Structures
     {
         [MarshalAs(UnmanagedType.ByValArray, SizeConst = 4)]
         public byte[] unknown_0; // 0x000 
-        public ulong id_anim; // 0x004 
-        public ulong id_actor; // 0x008 
+        public UInt32 id_anim; // 0x004 
+        public UInt32 id_actor; // 0x008 
         [MarshalAs(UnmanagedType.ByValArray, SizeConst = 308)]
         public byte[] unknown_C; // 0x00C 
     }
